lib/container: Only have one manifest fetch path
It was confusing to have a codepath that returns just the
digest, and another that fetches the parsed data and the digest.

Instead, fetch *unparsed* data (so that we don't need to expose
`oci::Manifest` as public API) as well as the digested image.

Prep for avoiding an extra round trip in the import path
to fetch the manifest when the caller already has it, in case
they wanted to do change detection.