lib/container: Convert to oci-archive, not docker-archive:
The use of `docker-archive:` originally came from
https://github.com/ostreedev/ostree-rs-ext/issues/18
but it turns out I was wrong, skopeo ends up staging everything
anyways.

In practice, what we really want to do here is land custom code
in skopeo to do what we need, instead of treating this as a `copy`
operation, because it isn't.

Relatedly, I also noticed in our `docker-archive:` parsing that we weren't
verifying the relationship of the layer symlink to the sha256 digest;
we were only verifying the *presence* of the symlink.
In practice, skopeo will already be doing the layer verification for
us, but this mechanism is clearer.

Prep for derivation.