store: Only close self-pipe when we're done
This took a crazy long time to debug but after lots of false
starts I think this is right. Basically what's going on is
we have async tasks that are talking over a `pipe()` inside
our own process.

We must not close the read side of the pipe until the
writer is done.

I believe this is dependent on tokio task scheduling order,
and it's way easier to reproduce when pinned to a single CPU.

Closes: https://github.com/ostreedev/ostree-rs-ext/issues/657
Signed-off-by: Colin Walters <walters@verbum.org>